requirements.txt
absl-py==1.2.0
asgiref==3.5.2
astunparse==1.6.3
backports.zoneinfo==0.2.1
black==22.10.0
cachetools==5.2.0
certifi==2022.9.24
charset-normalizer==2.1.1
cffi==1.15.1
click==8.1.3
colorama==0.4.6
cryptography==38.0.3
defusedxml==0.7.1
dj-rest-auth==2.2.5
Django==4.1.3
django-allauth==0.51.0
django-cors-headers==3.13.0
django-environ==0.9.0
django-resized==1.0.2
django-rest-passwordreset==1.3.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
idna==3.4
install==1.3.5
dlib==19.24.0
flatbuffers==22.9.24
gast==0.4.0
google-auth==2.12.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.49.1
h5py==3.7.0
idna==3.4
importlib-metadata==5.0.0
keras==2.10.0
Keras-Preprocessing==1.1.2
libclang==14.0.6
Markdown==3.4.1
MarkupSafe==2.1.1
numpy==1.23.3
opencv-contrib-python==4.6.0.66
opencv-python==4.6.0.66
opt-einsum==3.3.0
packaging==21.3
Pillow==9.3.0
protobuf==3.19.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyJWT==2.6.0
pyparsing==3.0.9
pytz==2022.6
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
scipy==1.9.2
six==1.16.0
sqlparse==0.4.3
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.10.0
tensorflow-estimator==2.10.0
tensorflow-io-gcs-filesystem==0.27.0
termcolor==2.0.1
typing_extensions==4.4.0
tzdata==2022.6
Werkzeug==2.2.2
wincertstore==0.2
wrapt==1.14.1
zipp==3.9.0
mypy-extensions==0.4.3
oauthlib==3.2.2
pathspec==0.10.2
platformdirs==2.5.4
pycparser==2.21
python3-openid==3.2.0
tomli==2.0.1
urllib3==1.26.13
